net/mlx5e: en_tc: Tunnels - decap on first chain hop, restore tunnel on miss
To allow header rewrite on inner, if metadata is enabled,
decap on first chain hop, and use register C to store the tunnel
info to match against instead, and restore it on miss.

If metadata is disabled, user can't decapasulate the packet with
goto action (otherwise if we then miss, the tunnel device won't be
able to get the decapsulated packet from PF).

Issue: 1806821
Change-Id: I088965f82c10ccba5a460b6718047575ffb055ba
Signed-off-by: Paul Blakey <paulb@mellanox.com>