net: Add tc chain extension to sk_buff
When offloading tc rules with goto chain actions such as
$ tc filter add ... chain 4 ... flower ... action goto chain 5
$ tc filter add ... chain 5 ... flower ip_proto tcp ... action goto chain 6
...

If there's a 'miss' in hardware (proto != tcp in the above example)
after jumping to some chain, we would like to fall back to software
and continue processing from where we left off (chain 5), as we might
have modified (and counted) the packet in earlier chains and can't redo
those in software.

To support this, introduce tc chain skb extension.
The extension will be set by drivers to specifiy from which tc
chain to start processing and in the next patch tc will use it to
start classifiction from that chain.

Issue: 1806821
Change-Id: Ibd14e99296e35e20222efa23307ac5ecaf313ceb
Signed-off-by: Paul Blakey <paulb@mellanox.com>
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>